# survey-parent/api-survey/src/main/resources/static/json/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "survey",
    "version": "v1.0",
    "title": "API文档",
    "termsOfService": "lego.cn",
    "contact": {
      "name": "lego",
      "url": "lego.cn",
      "email": "survey@legocloud.cn"
    }
  },
  "host": "192.168.101.197:50010",
  "basePath": "/",
...
  "paths": {
...
    "/survey/api-report/templateReport/create": {
      "post": {
        "tags": [
          "template-report-controller"
        ],
        "summary": "创建模板",
        "description": "创建模板",
        "operationId": "createUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "templateReport",
            "description": "templateReport",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TemplateReport"
            }
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-report/templateReport/delete": {
      "delete": {
        "tags": [
          "template-report-controller"
        ],
        "summary": "删除模板",
        "description": "删除模板",
        "operationId": "deleteUsingDELETE_11",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "id",
            "in": "query",
            "description": "模板id",
            "required": true,
            "type": "string",
            "x-example": "1"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-report/templateReport/list/{pageIndex}": {
      "get": {
        "tags": [
          "template-report-controller"
        ],
        "summary": "查询模板列表",
        "description": "查询模板列表",
        "operationId": "listUsingGET_12",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "pageIndex",
            "in": "path",
            "description": "请求页",
            "required": true,
            "type": "string",
            "default": "1",
            "x-example": "1"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": true,
            "type": "string",
            "default": "10",
            "x-example": "10"
          },
          {
            "name": "sectionCode",
            "in": "query",
            "description": "标段编号",
            "required": false,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO«RespDataVO«TemplateReport»»"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-report/templateReport/modify": {
      "put": {
        "tags": [
          "template-report-controller"
        ],
        "summary": "修改模板",
        "description": "修改模板",
        "operationId": "modifyUsingPUT_11",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "templateReport",
            "description": "templateReport",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TemplateReport"
            }
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "ApiError": {
      "type": "object",
      "title": "ApiError"
    },
...
    "RespDataVO«TemplateReport»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TemplateReport"
          }
        }
      },
      "title": "RespDataVO«TemplateReport»"
    },
...
    "RespVO": {
      "type": "object",
      "properties": {
        "info": {
          "type": "object"
        },
        "msg": {
          "type": "string"
        },
        "ret": {
          "type": "string"
        },
        "retCode": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "RespVO"
    },
...
    "RespVO«RespDataVO«TemplateReport»»": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/RespDataVO«TemplateReport»"
        },
        "msg": {
          "type": "string"
        },
        "ret": {
          "type": "string"
        },
        "retCode": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "RespVO«RespDataVO«TemplateReport»»"
    },
...
    "TemplateReport": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "createUser": {
          "type": "string",
          "description": "创建人"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 1,
          "description": "模板Id"
        },
        "name": {
          "type": "string",
          "description": "模板名称"
        },
        "path": {
          "type": "string",
          "description": "模板路径"
        },
        "reportNameRule": {
          "type": "string",
          "description": "生成的报告命名规则"
        },
        "scenes": {
          "type": "string",
          "description": "场景"
        },
        "sectionCode": {
          "type": "string",
          "description": "标段编码"
        },
        "type": {
          "type": "integer",
          "format": "int32",
          "example": 0,
          "description": "模板文件格式 0：EXCEL 1: WORD 2: TXT 3: PDF"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        }
      },
      "title": "TemplateReport"
    },
...
  }
}